backend/pyproject.toml
[tool.poetry]
name = "proofile-backend"
version = "0.1.0"
description = "Proofile Backend API"
authors = ["L. Singw <singwane.linda.m@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.118.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.43"}
alembic = "^1.16.5"
asyncpg = "^0.30.0"
redis = "^6.4.0"
celery = "^5.5.3"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {version = "==1.7.4", extras = ["bcrypt"]}
bcrypt = "==4.0.1"
python-multipart = "^0.0.20"
pydantic = "^2.11.9"
pydantic-settings = "^2.11.0"
openai = "^2.0.0"
anthropic = "^0.69.0" # Consider if this is a core dependency or for specific features
pinecone-client = "^6.0.0"
python-dotenv = "^1.1.1"
supabase = "^2.20.0"
email-validator = "^2.3.0"
python-magic = "^0.4.27"
aiofiles = "^23.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
httpx = "^0.28.1" # Moved from main dependencies
pytest-asyncio = "^0.23.7"
black = "^25.9.0"
ruff = "^0.13.2"
asgi-lifespan = "^2.1.0"
pytest-xdist = "^3.6.1"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"
